taskcoach/requirements.txt
PyPubSub
desktop3
wxPython
chardet
twisted
lockfile
